gruping svm implementations into one file
